\name{anno-package}
\alias{anno-package}
\alias{anno}
\docType{package}
\title{
\packageTitle{anno}
}
\description{
\packageDescription{anno}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{anno}
\packageIndices{anno}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{anno}

Maintainer: \packageMaintainer{anno}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
