% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annoRawToWide.r
\name{annoRawToWide}
\alias{annoRawToWide}
\title{Combines GEE formatted CSV file output and translates into wide format}
\usage{
annoRawToWide(rawP)
}
\arguments{
\item{rawP}{Path to raw GEE result files}
}
\description{
Combines GEE formatted CSV file output and translates into wide format
}
