% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepForGEE.r
\name{prepForGEE}
\alias{prepForGEE}
\title{Preps dataset for annotation in GEE.
1) Saves a copy of the data, but with anno_id added as a column.
2) Saves a shapefile formatted for upload to GEE.}
\usage{
prepForGEE(dat, datAnnoIdPF, geeShapefileP, ptsPerGrpReq = NULL,
  extraFields = NULL)
}
\arguments{
\item{dat}{Dataset}

\item{datAnnoIdPF}{Path and Filename specifying where to save original dataset, with anno_id added.}

\item{geeShapefileP}{Path of output shapefile, prepped for GEE. Final folder will be the name of the shapefile.}

\item{ptsPerGrpReq}{Requested number of points in each annotation group.}

\item{extraFields}{Fields, in addition to the default fields, that should be uploaded to GEE.}
}
\description{
Preps dataset for annotation in GEE.
1) Saves a copy of the data, but with anno_id added as a column.
2) Saves a shapefile formatted for upload to GEE.
}
