% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annoRawToWide.r
\name{annoRawToWide}
\alias{annoRawToWide}
\title{Combines GEE formatted CSV file output and translates into wide format}
\usage{
annoRawToWide(rawPath)
}
\arguments{
\item{rawPath}{\code{string} Path to raw GEE result files}
}
\description{
Combines GEE formatted CSV file output and translates into wide format
}
