% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepForGEE.r
\name{prepForGEE}
\alias{prepForGEE}
\title{Preps dataset for annotation in GEE.
1) Saves a copy of the data, but with anno_id added as a column.
2) Saves a shapefile formatted for upload to GEE.}
\usage{
prepForGEE(datPF, annoPF = NULL, shpDSN = NULL,
  ptsPerGrp = getOption("anno.ptsPerGrp"), extFields = NULL)
}
\arguments{
\item{datPF}{Path to dataset to annotate.}

\item{annoPF}{Path and Filename specifying where to save original dataset, with anno_id added.}

\item{shpDSN}{Path of output shapefile, prepped for GEE. Final folder will be the name of the shapefile.}

\item{ptsPerGrp}{Requested number of points in each annotation group.}

\item{extFields}{Fields, in addition to the default fields, that should be uploaded to GEE.}
}
\description{
Preps dataset for annotation in GEE.
1) Saves a copy of the data, but with anno_id added as a column.
2) Saves a shapefile formatted for upload to GEE.
}
