% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processGEEraw.r
\name{processGEEraw}
\alias{processGEEraw}
\title{High-level function that processes raw GEE results.
Converts to wide format (one variable per column).
Converts units if necessary.
Optionally joins to an existing file (file must have anno_id).}
\usage{
processGEEraw(rawP, joinToDat = NULL)
}
\arguments{
\item{rawP}{Path to raw GEE result files}

\item{joinToPF}{Path and filename of the file that the anno data should be joined to. Should have anno_id field.}
}
\description{
High-level function that processes raw GEE results.
Converts to wide format (one variable per column).
Converts units if necessary.
Optionally joins to an existing file (file must have anno_id).
}
